Split regions into separate file and also split the tests
