+++ src/index.js
@@ -1,8 +1,9 @@
 'use strict';
 
 const Client = require('./client');
+const regions = require('./regions');
 
 module.exports = {
   Client: Client,
-  regions: Client.regions,
+  regions: regions.regions,
 };
